moved delete buttons to edit component for ingredients component
