+++ react-app/src/components/Ingredients/AddIngredientsInput.module.css
@@ -23,4 +23,5 @@
 {
     display: flex;
     flex-direction: column;
+    margin-top: 15px;
 }
