+++ react-app/src/components/Ingredients/ShowIngredients.js
@@ -1,16 +1,7 @@
-import { useDispatch } from 'react-redux';
-import { deleteIngredient } from '../../store/recipe';
-
 import styles from './ShowIngredients.module.css';
 import styleUtils from '../RecipeUtils.module.css';
 
 const ShowIngredients = ({ setShowEdit, isOwner, recipeIngredients, recipeId }) => {
-    const dispatch = useDispatch();
-
-    const handleDelete = (e, id) => {
-        e.preventDefault();
-        dispatch(deleteIngredient(id, recipeId));
-    };
 
     return (
         <div className={styles.ingredients_inner_container}>
@@ -39,12 +30,6 @@ const ShowIngredients = ({ setShowEdit, isOwner, recipeIngredients, recipeId }) 
                             <div className={styles.ingredients_container}>
                                 { ingredient.ingredient }
                             </div>
-                            {isOwner &&
-                                <i
-                                className={`fas fa-minus-circle ${styleUtils.delete_item}`}
-                                onClick={(e) => handleDelete(e, ingredient.id)}
-                                ></i>
-                            }
                         </div>
                     ))
                 }
